app/requirements.txt
fastapi
uvicorn[standard]
openslide-python